+++ webpack.config.js
@@ -3,11 +3,14 @@
 
 const path = require('path')
 const webpack = require('webpack')
+const {BundleAnalyzerPlugin} = require('webpack-bundle-analyzer')
+const prod = process.env.NODE_ENV === 'production'
 
 module.exports = {
+  mode: prod ? 'production' : 'development',
   entry: [path.join(__dirname, 'src', 'react-markdown.js')],
   output: {
-    library: 'reactMarkdown',
+    library: 'ReactMarkdown',
     libraryTarget: 'umd',
     path: path.join(__dirname, 'umd'),
     filename: 'react-markdown.js'
@@ -21,10 +24,13 @@ module.exports = {
     }
   },
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.js$/,
-        loader: 'babel-loader'
+        exclude: /node_modules/,
+        use: {
+          loader: 'babel-loader'
+        }
       }
     ]
   },
@@ -32,6 +38,6 @@ module.exports = {
     new webpack.DefinePlugin({
       'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV || 'development')
     }),
-    new webpack.optimize.UglifyJsPlugin()
-  ]
+    process.env.ANALYZE_BUNDLE && new BundleAnalyzerPlugin()
+  ].filter(Boolean)
 }